xstatic-package for Bootswatch Bootstrap Themes
Bootswatch provides many out of the box Bootstrap themes.
We plan on using this in Horizon to provide a theme that
implements Google's Material Design.

We plan on using this in Horizon to provide a theme that
implements Google's Material Design.

The 'material' theme gives an example of how to make use of a 3rd party theme
using the theming functionality.  In addition to incorporating a theme,
loaded as a static asset using requirements.txt, it also gives examples of how
to cleanly override styles, variables, icon fonts and Django templates.

This theme is replacing 'blue' as the example of how to use a theme other than
'default'.

Partially-Implements: blueprint horizon-theme-css-reorg
Change-Id: I83753ca7f50d8e6d3051af4a7a929665e0de4d4d